src/requirements.txt
torch
joblib
pandas
numpy
scikit-learn
matplotlib
seaborn
Pillow
gradio